edge/requirements.txt
pip~=21.3.1
wheel~=0.37.1
paho-mqtt~=2.1.0
pika~=1.3.2
requests~=2.32.3